Move ModerationIdentifier Doctrine type
It was misplaced in the DoctrineEntities namespace, but should be in the
DoctrineTypes.
